buffer: add advance(unsigned) back
advance(unsigned) and seek(int) were replaced with advance(size_t> and
seek(ssize_t) in 053bfa6. this change broke the backward compatibility
of librados/librbd's ABI.

- deprecates advance(unsigned) and seek(int)
- update the callers to use the prefered versions.

Fixes: http://tracker.ceph.com/issues/17809

+++ src/include/buffer.h
@@ -390,8 +390,8 @@ namespace buffer CEPH_BUFFER_API {
 	//return off == bl->length();
       }
 
-      void advance(ssize_t o);
-      void seek(size_t o);
+      void advance(int o);
+      void seek(unsigned o);
       char operator*() const;
       iterator_impl& operator++();
       ptr get_current_ptr() const;
@@ -434,8 +434,8 @@ namespace buffer CEPH_BUFFER_API {
       iterator(bl_t *l, unsigned o=0);
       iterator(bl_t *l, unsigned o, list_iter_t ip, unsigned po);
 
-      void advance(ssize_t o);
-      void seek(size_t o);
+      void advance(int o);
+      void seek(unsigned o);
       char operator*();
       iterator& operator++();
       ptr get_current_ptr();
